9.1.12 Добавит возврат результата по условию
Если интернета нет, мы сразу возвращаем разрешенный промис с данными.

Для запроса всех данных это пока пустой массив.

Для остальных же методов мы возвращаем

+++ src/api/provider.js
@@ -1,3 +1,6 @@
+import nanoid from "nanoid";
+import Task from "../models/task";
+
 export default class Provider {
   constructor(api, store) {
     this._api = api;
@@ -5,19 +8,40 @@ export default class Provider {
   }
 
   getTasks() {
-    return this._api.getTasks();
+    if (this._isOnLine()) {
+      return this._api.getTasks();
+    }
+
+    return Promise.resolve(Task.parseTasks([]));
   }
 
   createTask(task) {
-    return this._api.createTask(task);
+    if (this._isOnLine()) {
+      return this._api.createTask(task);
+    }
+
+    // Нюанс в том, что при создании мы не указываем id задачи, нам его в ответе присылает сервер.
+    // Но на случай временного хранения мы должны позаботиться и о временном id
+    const fakeNewTaskId = nanoid();
+    const fakeNewTask = Task.parseTask(Object.assign({}, task.toRAW(), {id: fakeNewTaskId}));
+
+    return Promise.resolve(fakeNewTask);
   }
 
   updateTask(id, task) {
-    return this._api.updateTask(id, task);
+    if (this._isOnLine()) {
+      return this._api.updateTask(id, task);
+    }
+
+    return Promise.resolve(task);
   }
 
   deleteTask(id) {
-    return this._api.deleteTask(id);
+    if (this._isOnLine()) {
+      return this._api.deleteTask(id);
+    }
+
+    return Promise.resolve();
   }
 
   _isOnLine() {
